project/requirements.txt
fastapi
uvicorn
streamlit
requests
groq
python-dotenv
reportlab